Split CLI tests into basic usage and integration tests.
Change units from Int to ByteSize for byte-sized-limited, update comments and json serdes, add byte size subtraction, update tests.

Make some const values core package visible.

Make memory option byte-sizes.

Add toString on byte-size and update error messages.

Remove printlns.

Update and move tests around between packages.

Make TimeLimit constructor accept only a duration.

Revert change to log/memory limit constructor so that the type is int not bytesize (really would want to constrain this to be MB but the type conversions aren't that rich and arguably not worth it as this can be hidden behind the private constructor).

Change Bytes to B in toString to avoid 1 bytes vs 1 byte. Should it be byte[s] and append the s only when necessary?

Emit parameters should test fail.

Use python CLI for http test as go cli assumes https.

Add some more tests for limits.

Use Python CLI for two tests that do not work with Go CLI.

Add retry on list operations which might need some time to be consistent.
